backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart
aiofiles
openai==1.51.2
moviepy==1.0.3
opencv-python==4.10.0.84
face_recognition==1.3.0
textblob
python-dotenv
numpy==2.1.1
pydantic==2.9.2
requests
filelock